Change writable stream error semantics to be more uniform
The main motivation for this change is to centralize the erroring of a writable stream into a single abstract operation, WritableStreamError(), which takes care of all the relevant operations together, instead of sprinkling them through separate state transitions. This closes #704 by subsuming it; now ResetQueue() is always called when erroring the stream, no matter the code path.

The effort to do this revealed a semantic mismatch wherein we deciding abort()ing during an in-flight close should error the stream, but the abort should suceed. This was not a good decision, in retrospect; it makes this particular instance of erroring the stream very unusual. And we recognized that it was weird, because we made abort() return a fulfilled promise. Instead we should treat this as a success and close the stream, reflected in both the return value of abort() and in the stream's state.

This also introduces internal methods to writable stream controllers, paralleling the structure already in place for readable stream controllers. These methods were already there, disguised as abstract operations; since this change introduces a third such hook, it was time to properly pull them out as internal methods.