Update WebSocket examples
I recently learned that WebSockets will tell you if they closed cleanly. This allows us to illustrate returning a potentially-rejected promise from the underlying sink's close() and abort() methods.

This also makes the writable streams error themselves if the WebSockets are closed by the server (as opposed to closed by the client), as this represents an unexpected cessation of the ability to write data.

Finally this synchronizes the examples so that they both send a status code of 1000 for closing successfully.