Revise abstract, introduction, and model
Making the abstract shorter fixes #769. The introduction incorporates some of its former content, and also was updated in light of how streams and their usage have evolved over time. Finally, it also includes a brief bit about web-developer-created streams, which was curiously absent.

The model section was updated to tie the abstract concepts there to concrete JavaScript classes. Back when it was written, there was some idea of using duck-typing, but that has been abandoned, and this is much more helpful.

This also marks up non-normative sections as non-normative in the source, which allows Bikeshed to properly compute normative vs. non-normative references and detect misuses of RFC 2119 keywords. A few such misuses were corrected.