Remove TransformStream's writableDone member
The [[writableDone]] slot was only used to avoid calling
WritableStreamDefaultControllerError when the stream wasn't writable. This is
exactly the semantics that WritableStreamDefaultControllerErrorIfNeeded has, so
use that instead.

If abort() is called on the writable then cancel() is called on the readable(),
the reason passed to abort() should be the one that errors the writable. Add
tests to verify this behaviour.

Fix a bug where a call to the underlying sink abort() method after the
TransformStream was errored would cause an assert.