Permit any pending writes to complete on error
Previously if the write() or start() methods of TransformStreamDefaultSink were
waiting for backpressure to be relieved when the readable was errored, then they
would never complete, meaning that things depending on them such as the promises
returned by writer.write() and writer.close() would never resolve either.

Set backpressure to false when erroring the TransformStream to permit pending
operations to complete.

Also add tests for these cases.