Permit closing the writable to close the readable, even with HWM 0
Previously a call to writer.close() on a TransformStream with
readableStrategy.highWaterMark 0 and no chunks queued would not result
in the readable side being closed. This was because the backpressure
would mean that the promises returned by start() or write() would never
settle.

Instead of waiting for backpressure to clear at the end of every call to
sink start() and write(), instead wait for it to clear at the start of a
call to sink write(), before calling transformer.transform().

When there is no backpressure, write() will still call transform()
synchronously.

The results of only one test changed as a result of this. It has been
updated. A new test has been added to verify that writer.close() now
results in the readable being closed with HWM 0.