Remove TransformStream _readableClosed member
Instead inspect the state of the ReadableStream directly using a new abstract
operation, ReadableStreamDefaultControllerCanCloseOrEnqueue operation. This new
operation is also used by the ReadableStream implementation in place of the
equivalent comparisons.

This avoids redundant state tracking.

Also allow erroring a TransformStream after close is requested

Previously, TransformStream would not permit erroring the readable side after
controller.close() was called. Change it to be consistent with ReadableStream so
that erroring is still possible as long as all enqueued chunks have not yet been
read.

Also add tests for the new behaviour.