Fix interaction between writable.abort() and readable.cancel()
If writable.abort() is called when the WritableStream is waiting for
another operation to complete, and then readable.cancel() is called
before the queued abort is processed, an assert() would happen due to
trying to error the TransformStream twice.

Change TransformStreamDefaultSink's abort() method to use
TransformStreamErrorIfNeeded to correctly handle this case. Also add
tests for this case, and to confirm that it is the call to abort() that
errors the writable in this case.

Also remove the [[writableDone]] slot. It was only used to avoid calling
WritableStreamDefaultControllerError when the stream wasn't
writable. This is exactly the semantics that
WritableStreamDefaultControllerErrorIfNeeded has, so use that instead.