Add TransformStream tests for re-entrant strategies
Add tests to ensure that implementations of TransformStream behave correctly
when operations are triggered from within the readableStrategy.size function.

This is not really a "supported" use case as any code that does this is either
malicious or broken, but implementations need to avoid crashes or memory
corruption in these cases.

Also merge changes to recording-streams.js from web-platform-tests that are
required to make these tests work.