Remove _errored and _storedError from TransformStream
These slots are mostly redundant with the same state on the readable
side, so use the ReadableStream [[state]] and [[storedError]] instead to
reduce complexity.

It is possible for the readable to be "closed" where previously _errored
was set, but this doesn't make any practical difference as the
implementation threw a TypeError in both cases anyway.

TransformStreamError is renamed to TransformStreamDefaultControllerError to
better reflect its new semantics.

Closes #796.