Detect if the TransformStream errored inside write() and reject
If a TransformStream errored while the TransformStreamDefaultSink write() method
was waiting for backpressure to be relieved, it would assert inside
the TransformStreamStream operation.

Since it is not useful to call transform() once the stream is errored, return a
rejection from the write() method instead.

Also add a test for this case.