Remove assertion that RSDefaultControllerClose doesn't throw
ReadableStreamDefaultControllerClose is defined in the standard as "nothrow" but
TransformStreamCloseReadableInternal contains an assert that it doesn't
throw. The assertion that it doesn't throw is already implied by the "nothrow"
annotation. Since it is good to aim for a 1:1 relationship between the reference
implementation and the standard text, the reference implementation shouldn't
contain this assert either.

Removed.