Remove the check of the resolved value of _backpressurePromise
TransformStreamSetBackpressure would resolve _backpressurePromise with the
new value of `backpressure`, then a fulfillment handler would verify that this
is not equal to the old value. Since this is already guaranteed by the other
logic inside of TransformStreamSetBackpressure the benefit of this check is
small. On the other hand, the cost, in terms of having an additional fulfillment
handler, is high.

Remove this redundant sanity check.