Use "debug" package for logging
Use the "debug" package https://www.npmjs.com/package/debug for logging.

Debugging can be enabled by setting the DEBUG environment variable, for example

  DEBUG=streams:* npm test

This change converts existed commented-out log statements in
transform-streams.js to use the debug package. No new logging statements are
included in this change.

It is anticipated that new logging statements will be added as and when a need
for them is found.

Closes #499.