Remove legacy test infrastructure and abstract-ops tests
Over time, we've moved all tests over to the web-platform-tests format, either upstream or in this repository's to-upstream-wpts folder. The sole remaining exception was tests for the IsReadableStreamDisturbed abstract operation, which is not web-exposed, but instead called by other specifications.

This commit removes the tests for IsReadableStreamDisturbed; they can be replaced by web platform tests for the web-exposed effects (see https://github.com/w3c/web-platform-tests/pull/7630). This allows us to remove all the infrastructure around these legacy tests, and concentrate solely on web-platform-tests.

While here, we update some of the dependencies for our test tooling.