Initialise readableStrategy without using an object literal
The strategy argument passed to the ReadableStream constructor within the
TransformStream argument used an object literal. The standard text only
uses object literal syntax in default arguments. In order to align the
reference implementation with the standard text, create the
`readableStrategy` object in explicit algorithm steps before passing it
to the constructor.

No functional changes.