Add note about initialisation of [[backpressure]] slot
Explain why the constructor initialises the [[backpressure]] slot
to *undefined* and note that implementations can do something else.

Also add a note that TransformStreamError can be used even when an error
has already occurred.

Also change TransformStreamDefaultControllerEnqueue to call
TransformStreamErrorWritableAndUnblockWrite rather than
TransformStreamError when ReadableStreamDefaultControllerEnqueue throws,
since we know that the readable side will already be errored in that
case.