Add 2 operations CreateAlgorithmWith...FromUnderlyingMethod
Add CreateAlgorithmWithNoParametersFromUnderlyingMethod and
CreateAlgorithmWithOneParameterFromUnderlyingMethod. These both either create an
algorithm that calls the method or does nothing depending on whether or not the
method is defined or not. They can both accept additional arguments to pass to
the method. The difference between the two versions is how many parameters the
created algorithm should take.

All of ReadableStream, TransformStream and WritableStream have been modified to
use the new operations. This involves creating the controller object in the
caller of Setup...Stream...Controller() so that it can be passed as an argument
to the CreateAlgorithm methods.