Suppress Bikeshed warnings about <vars> used once
The arguments to methods and abstract operations are marked up with
Bikeshed <var> tags, but the actual algorithms are marked up using
emu-algify. Bikeshed does not recognize the uses of the arguments within
the emu-algify blocks, and so warns that the <var>s were only used once.

Use the "ignore" parameter on affected <var> tags to suppress the
warning. It is only used on <var> tags that actually trigger the
warning, as the benefit of having the other <var> tags checked outweighs
the potential benefit of consistency.

No functional or visible changes.

Fixes #869.