Do not throw when calling RS's controller.error()
Previously, we would throw an exception when calling controller.error() for a readable stream that was not readable. As noted in #821, this is inconsistent with the behavior introduced for writable and transform streams. The fact that the spec is constantly checking whether it's in the right state, e.g. via ReadableStreamDefaultControllerCallPullIfNeeded, is also a strong indicator that we should not throw.

This simplifies various abstract operations by centralizing state checks and removing ReadableStreamDefaultControllerCallPullIfNeeded.

Closes #821.