Revise property access order
Make the order that properties are accessed on the underlying
source, underlying sink, transformer and strategy objects more
consistent. Also make the order that properties are validated more
consistent.

The general order of lookups is now size, highWaterMark, type,
methods. In general values are validated immediately after lookup, but
size and highWaterMark are exceptions. They are always looked up, but
the validation needed depends on the value of "type", so lookup and
validation are separated. This applies even to WritableStream and
TransformStream, to allow for the addition of byte types to those in
future.

The order of lookups in the reference implementation now matches the
standard.

Also convert autoAllocateChunkSize to a number before checking whether
it is integer, for consistency with the way other numeric input is
handled.

Fixes: #917, #921