Changes to Example 8.9: Fix code issue, add const, add text decode step (#924)
In a real world scenario, the response body is encoded as a Uint8Array.
As the LipFuzzTransformer requires string chunks in order to work,
the Uint8Array chunks need to be decoded to strings first. To clear up
any ambiguity for the reader, a text decoding step is added before the
LipFuzzTransformer is applied so the example works in browsers.