ShouldCallPull: Assert desiredSize is not null
ReadableStreamDefaultControllerShouldCallPull implicitly assumes that
desiredSize is not null when it compares it to 0. This is guaranteed by
the check of
ReadableStreamDefaultControllerCanCloseOrEnqueue(controller), but it's
not obvious. Add an assert to make it explicit.

Also do the same for ReadableByteStreamControllerShouldCallPull.

No functional changes.

Also, ReadableStreamDefaultControllerShouldCallPull was missing an exclamation
mark before ReadableStreamDefaultControllerGetDesiredSize. Fix it.