Add a text encoding step to example 8.9
In #924, a text decoding step was added to example 8.9 so it could run in the browser. However, the reverse needs to happen in order to construct a valid Response with the transformed stream. Specifically, in step 3 of "read all bytes from a ReadableStream" from the Fetch specification, the stream must produce Uint8Array objects for it to be consumed correctly.