Set shuttingDown at the beginning of pipe shutdown
In the implementation of pipeTo(), setting _shuttingDown_ after flushing
unwritten reads makes it difficult for implementations to obey the
requirement that "Shutdown must stop activity". Generally it means they
have to track separately whether they've started to shutdown, making the
_shuttingDown_ variable redundant. The exact behaviour while unwritten
reads are being flushed may also be ambiguous.

Change to set _shuttingDown_ at the beginning of the _Shutdown with an
action_ and _Shutdown_ steps.

No changes to reference implementation as it already works this way.

Fixes #934.