Make rejections from transformAlgorithm error the readable
Currently rejections from transformAlgorithm error the writable side of
a transform stream but don't error the readable side. This is an
unexpected difference from the behaviour of transformer.transform.

Make the behaviour consistent so a rejection from transformAlgorithm
errors both sides.

This only affects other standards which create TransformStream objects.
The change is not visible to user JavaScript or web platform tests.

Fixes #946.