Update the standard text to match the ref impl
Reflect the fixes to the algorithms made in the reference implementation
back into the standard.

In particular, in order to correctly handle signals that are already
aborted, _abortAlgorithm_ is defined earlier so that it can be called
synchronously if necessary, and calls to WritableStreamAbort and
ReadableStreamCancel are guarded by state checks.