Give pipeTo more latitude about byob chunk sizes
If pipeTo() uses a BYOB reader, it is required to use
WritableStreamDefaultWriterGetDesiredSize(writer) to determine chunk
sizes. However, if, for example, the HWM is too low, the resulting value
may be an unsuitable chunk size. Relax the language to give
implementations latitude to use other information to choose an
appropriate chunk size.

Closes #938.