In ReadableStreamTee, replace _closedOrErrored_ with _closed_
DefaultReaderRead will never resolve after the [[closedPromise]] has rejected,
therefore setting _closedOrErrored_ on rejection makes no difference. Rename the
variable to just _closed_ and only use it to signal that a read has returned
with "done" set to true.

Also change the lookups of "done" and "value" in ReadableStreamTee from "? Get"
to "! Get" as they can never throw, and pullAlgorithm returning an abrupt
completion is not supported anyway.

Rearrange the steps slightly to reduce redundant work.

Closes #991.