ReadableStreamTee: do not read when already pulling
Currently, when one branch calls pull(), it is possible that we call
read() - even though a read() is still pending from a previous pull()
call by the other branch. This can lead to reading too many chunks from
the original streams, and unnecessarily filling up the queues of the
branches past their high water mark.

This PR fixes that issue by only starting a new read() if the previous
read() was resolved. As a bonus, there will now be only one read() that
resolves with { done: true } when the original stream closes, so we no
longer need to keep track of whether we already closed the two branches.

Fixes #996