Reference implementation: fix pipe not aborting
When both the preventCancel and preventAbort flags are set, the abortAlgorithm passes an empty list of actions to "wait for all". This helper should immediately resolve the returned promise when no promises are given, but instead the promise remained pending forever.

This fixes the reference implementation of the "wait for all" helper to resolve correctly when given an empty list of promises. This matches the spec change in https://github.com/w3ctag/promises-guide/pull/58.

Originally reported in https://github.com/whatwg/streams/issues/1004#issuecomment-511037767.