Settle reader.closed promise before closing/erroring read requests
Async iterators, because of their use of (sync) close/error steps on request requests, could trigger a strange case where some of the stream's internal state is temporarily out of sync. This would manifest in it trying to resolve an already-resolved promise, which is allowed, but is a sign that something weird is going on in the spec.

After this change, the stream does all of its own internal state updates, before calling the close/error steps. Then, those steps can observe and manipulate the "final" state of the stream. Thus, when the async iterator releases its lock, it will replace the [[closedPromise]] with a newly rejected promise instead.

This is a small but observable change for both web authors and other specifications:

* For web authors, reader.closed will resolve with undefined before reader.read() resolves with { done: true, value: undefined }.

* For other specifications using the "read a chunk" algorithm, when their close steps are run, they will now observe that reader.closed is already resolved (see comment on https://github.com/whatwg/streams/issues/1100#issuecomment-761282151).

Closes #1100.