Support teeing readable byte streams
Teeing a readable byte stream will now create two readable byte streams (instead of two "default" readable streams). Fixes #1111.

This implements the "hard" solution mentioned there: when a branch is being read using a BYOB reader, the view of that read request is forwarded to the original stream, so it can read directly into that view. After the view is filled, an identical copy is created and enqueued on the other branch, so they both see the same data.